Extract common `BoundedFieldGenerator` class
